continue simplifying types.ts
