+++ src/command/render/codetools.ts
@@ -8,7 +8,8 @@
 import { Document, Element } from "deno_dom/deno-dom-wasm.ts";
 
 import { kCodeTools, kKeepSource } from "../../config/constants.ts";
-import { Format, isHtmlOutput } from "../../config/format.ts";
+import { Format } from "../../config/types.ts";
+import { isHtmlOutput } from "../../config/format.ts";
 import { executionEngineCanKeepSource } from "../../execute/engine.ts";
 import { ExecutionEngine, ExecutionTarget } from "../../execute/types.ts";
 import { formatHasBootstrap } from "../../format/html/format-html-bootstrap.ts";
