+++ src/command/render/crossref.ts
@@ -11,7 +11,7 @@ import {
   kNumberOffset,
   kNumberSections,
 } from "../../config/constants.ts";
-import { FormatPandoc } from "../../config/format.ts";
+import { FormatPandoc } from "../../config/types.ts";
 import { PandocFlags } from "../../config/flags.ts";
 import { Metadata } from "../../config/types.ts";
