+++ src/command/render/output.ts
@@ -26,7 +26,8 @@ import {
   kTemplate,
   kVariant,
 } from "../../config/constants.ts";
-import { Format, isHtmlOutput } from "../../config/format.ts";
+import { Format } from "../../config/types.ts";
+import { isHtmlOutput } from "../../config/format.ts";
 
 import {
   quartoLatexmkOutputRecipe,
