+++ src/config/toc.ts
@@ -7,7 +7,7 @@
 
 import { kTableOfContents, kToc, kTocTitle } from "./constants.ts";
 import { PandocFlags } from "./flags.ts";
-import { Format } from "./format.ts";
+import { Format } from "./types.ts";
 
 export const kTocFloat = "toc-float";
