+++ src/config/types.ts
@@ -4,7 +4,262 @@
 * Copyright (C) 2020 by RStudio, PBC
 *
 */
+import { Document } from "deno_dom/deno-dom-wasm.ts";
+
+import {
+  kAtxHeaders,
+  kCache,
+  kCiteMethod,
+  kCiteproc,
+  kCodeFold,
+  kCodeLink,
+  kCodeSummary,
+  kCodeTools,
+  kCss,
+  kEcho,
+  kEngine,
+  kEPubCoverImage,
+  kEval,
+  kExecuteDaemon,
+  kExecuteDaemonRestart,
+  kExecuteDebug,
+  kExecuteEnabled,
+  kExecuteIpynb,
+  kFigAlign,
+  kFigDpi,
+  kFilterParams,
+  kFilters,
+  kFreeze,
+  kHighlightStyle,
+  kHtmlMathMethod,
+  kInclude,
+  kIncludeAfterBody,
+  kIncludeBeforeBody,
+  kIncludeInHeader,
+  kKeepHidden,
+  kKeepIpynb,
+  kKeepYaml,
+  kLatexAutoInstall,
+  kLatexAutoMk,
+  kLatexClean,
+  kLatexMakeIndex,
+  kLatexMakeIndexOpts,
+  kLatexMaxRuns,
+  kLatexMinRuns,
+  kLatexOutputDir,
+  kLatexTlmgrOpts,
+  kListings,
+  kMarkdownHeadings,
+  kMergeIncludes,
+  kNumberOffset,
+  kNumberSections,
+  kOutput,
+  kOutputFile,
+  kPdfEngine,
+  kPdfEngineOpt,
+  kPdfEngineOpts,
+  kPreferHtml,
+  kSectionDivs,
+  kSelfContained,
+  kSelfContainedMath,
+  kShiftHeadingLevelBy,
+  kTableOfContents,
+  kTemplate,
+  kToc,
+  kTocTitle,
+  kTopLevelDivision,
+  kVariables,
+  kVariant,
+  kWarning,
+} from "../config/constants.ts";
+
+import {
+  kError,
+  kFigFormat,
+  kFigHeight,
+  kFigWidth,
+  kKeepMd,
+  kKeepSource,
+  kKeepTex,
+  kOutputDivs,
+  kOutputExt,
+  kPageWidth,
+} from "./constants.ts";
+
+import { PandocFlags } from "./flags.ts";
+
+export const kDependencies = "dependencies";
+export const kSassBundles = "sass-bundles";
+export const kHtmlPostprocessors = "html-postprocessors";
+export const kBodyEnvelope = "body-envelope";
+export const kTextHighlightingMode = "text-highlighting-mode";
+export const kQuartoCssVariables = "css-variables";
 
 export type Metadata = {
   [key: string]: unknown;
 };
+
+export interface FormatDependency {
+  name: string;
+  version?: string;
+  meta?: Record<string, string>;
+  links?: { rel: string; href: string }[];
+  scripts?: DependencyFile[];
+  stylesheets?: DependencyFile[];
+  resources?: DependencyFile[];
+}
+
+export interface DependencyFile {
+  name: string;
+  path: string;
+  attribs?: Record<string, string>;
+}
+
+export interface BodyEnvelope {
+  header?: string;
+  before?: string;
+  after?: string;
+}
+
+export interface SassLayer {
+  use?: string[];
+  defaults: string;
+  functions: string;
+  mixins: string;
+  rules: string;
+}
+
+export interface SassBundle {
+  key: string;
+  dependency: string;
+  user?: SassLayer;
+  quarto?: SassLayer;
+  framework?: SassLayer;
+  loadPath?: string;
+}
+
+export interface FormatExtras {
+  pandoc?: FormatPandoc;
+  metadata?: Metadata;
+  [kTocTitle]?: string;
+  [kIncludeInHeader]?: string[];
+  [kIncludeBeforeBody]?: string[];
+  [kIncludeAfterBody]?: string[];
+  [kFilters]?: {
+    pre?: string[];
+    post?: string[];
+  };
+  [kFilterParams]?: Record<string, unknown>;
+  html?: {
+    [kDependencies]?: FormatDependency[];
+    [kSassBundles]?: SassBundle[];
+    [kBodyEnvelope]?: BodyEnvelope;
+    [kHtmlPostprocessors]?: Array<(doc: Document) => Promise<string[]>>;
+    [kTextHighlightingMode]?: "dark" | "light" | undefined;
+    [kQuartoCssVariables]?: string[];
+  };
+}
+
+// pandoc output format
+export interface Format {
+  render: FormatRender;
+  execute: FormatExecute;
+  pandoc: FormatPandoc;
+  metadata: Metadata;
+  formatExtras?: (
+    input: string,
+    flags: PandocFlags,
+    format: Format,
+  ) => Promise<FormatExtras>;
+  extensions?: Record<string, unknown>;
+}
+
+export interface FormatRender {
+  [kKeepTex]?: boolean;
+  [kKeepYaml]?: boolean;
+  [kKeepSource]?: boolean;
+  [kKeepHidden]?: boolean;
+  [kPreferHtml]?: boolean;
+  [kOutputDivs]?: boolean;
+  [kVariant]?: string;
+  [kOutputExt]?: string;
+  [kPageWidth]?: number;
+  [kFigAlign]?: "left" | "right" | "center" | "default";
+  [kCodeFold]?: "none" | "show" | "hide" | boolean;
+  [kCodeSummary]?: string;
+  [kCodeLink]?: boolean;
+  [kCodeTools]?: boolean | {
+    source?: boolean;
+    toggle?: boolean;
+    caption?: string;
+  };
+  [kMergeIncludes]?: boolean;
+  [kSelfContainedMath]?: boolean;
+  [kLatexAutoMk]?: boolean;
+  [kLatexAutoInstall]?: boolean;
+  [kLatexMinRuns]?: number;
+  [kLatexMaxRuns]?: number;
+  [kLatexClean]?: boolean;
+  [kLatexMakeIndex]?: string;
+  [kLatexMakeIndexOpts]?: string[];
+  [kLatexTlmgrOpts]?: string[];
+  [kLatexOutputDir]?: string | null;
+}
+
+export interface FormatExecute {
+  // done
+  [kFigWidth]?: number;
+  [kFigHeight]?: number;
+  [kFigFormat]?: "retina" | "png" | "jpeg" | "svg" | "pdf";
+  [kFigDpi]?: number;
+  [kCache]?: true | false | "refresh" | null;
+  [kFreeze]?: true | false | "auto";
+  [kExecuteEnabled]?: true | false | null;
+  [kExecuteIpynb]?: true | false | null;
+  [kExecuteDaemon]?: number | boolean | null;
+  [kExecuteDaemonRestart]?: boolean;
+  [kExecuteDebug]?: boolean;
+  [kEngine]?: string;
+  [kEval]?: true | false | null;
+  [kError]?: boolean;
+  [kEcho]?: boolean;
+  [kOutput]?: boolean;
+  [kWarning]?: boolean;
+  [kInclude]?: boolean;
+  [kKeepMd]?: boolean;
+  [kKeepIpynb]?: boolean;
+}
+
+export interface FormatPandoc {
+  from?: string;
+  to?: string;
+  writer?: string;
+  [kTemplate]?: string;
+  [kOutputFile]?: string;
+  standalone?: boolean;
+  [kSelfContained]?: boolean;
+  [kVariables]?: { [key: string]: unknown };
+  [kAtxHeaders]?: boolean;
+  [kMarkdownHeadings]?: boolean;
+  [kIncludeBeforeBody]?: string[];
+  [kIncludeAfterBody]?: string[];
+  [kIncludeInHeader]?: string[];
+  [kCiteproc]?: boolean;
+  [kCiteMethod]?: string;
+  [kFilters]?: string[];
+  [kPdfEngine]?: string;
+  [kPdfEngineOpts]?: string[];
+  [kPdfEngineOpt]?: string;
+  [kEPubCoverImage]?: string;
+  [kCss]?: string | string[];
+  [kToc]?: boolean;
+  [kTableOfContents]?: boolean;
+  [kListings]?: boolean;
+  [kNumberSections]?: boolean;
+  [kNumberOffset]?: number[];
+  [kHighlightStyle]?: string;
+  [kSectionDivs]?: boolean;
+  [kHtmlMathMethod]?: string | { method: string; url: string };
+  [kTopLevelDivision]?: string;
+  [kShiftHeadingLevelBy]?: number;
+}
