+++ src/project/project-index.ts
@@ -13,7 +13,7 @@ import { dirAndStem, pathWithForwardSlashes } from "../core/path.ts";
 import { PartitionedMarkdown } from "../core/pandoc/pandoc-partition.ts";
 
 import { Metadata } from "../config/types.ts";
-import { Format } from "../config/format.ts";
+import { Format } from "../config/types.ts";
 
 import { kOutputFile, kTitle } from "../config/constants.ts";
