+++ src/project/types/book/book-chapters.ts
@@ -15,7 +15,7 @@ import {
   kTitle,
   kToc,
 } from "../../../config/constants.ts";
-import { Format } from "../../../config/format.ts";
+import { Format } from "../../../config/types.ts";
 import { Metadata } from "../../../config/types.ts";
 
 import { ProjectConfig, ProjectContext } from "../../project-shared.ts";
