+++ src/project/types/website/website-navigation.ts
@@ -28,7 +28,7 @@ import {
   kHtmlPostprocessors,
   kSassBundles,
   SassBundle,
-} from "../../../config/format.ts";
+} from "../../../config/types.ts";
 import { PandocFlags } from "../../../config/flags.ts";
 import {
   hasTableOfContents,
